requirements.txt.txt
streamlit==1.42.0
yfinance
google-generativeai
python-dotenv